imgui windows now stored in map instead of lots of bools
This was mostly done in order to clean up renderAllTheThings() method in
ChessDrawer. Now instead of an if statement for every bool per window,
I can use a range based for loop to iterate over the std::unorderd_map
of open windows. The map maps from an enum called WindowTypes to
member function pointers and a bool. The bool portion is used for
marking that map element for defered deletion. This is easier than
trying to erase while looping over the container.