added a connection manager and a lower level TCP server connection
The ConnectionManager class now is responsible for managing the lower
level ServerConnection class. The ConnectionManager gets messages from
ServerConnection and then builds them and posts the relevant event with
the new event system. The ConnectionManager also builds messages to give
to the ServerConnection. This is all subject to change as of this
commit. The code does not compile yet.